LPS-62396 refactor and move logic out of action
